backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg[binary]>=3.2.0
pgvector>=0.3.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sse-starlette>=2.1.0
pypdf>=4.3.0
python-docx>=1.1.2

